# public/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Quorum Key Manager API server documentation.",
        "title": "Quorum Key Manager API",
        "contact": {},
        "license": {
            "name": "Business Source License 1.1",
            "url": "https://mariadb.com/bsl11/"
        },
        "version": "0.1"
    },
    "host": "https://consensysquorum.github.io/quorum-key-manager",
    "basePath": "/",
    "paths": {
...
        "/stores/{storeName}/keys": {
            "get": {
                "description": "List identifiers of keys store on selected Store",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keys"
                ],
                "summary": "List Key ids",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Store Identifier",
                        "name": "storeName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of key ids",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/types.KeyResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stores/{storeName}/keys/verify-signature": {
            "post": {
                "description": "Verify if signature data was signed by a specific key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keys"
                ],
                "summary": "Verify key signature",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Store Identifier",
                        "name": "storeName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Key identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful verification"
                    },
                    "422": {
                        "description": "Cannot verify signature",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stores/{storeName}/keys/{id}": {
            "get": {
                "description": "Retrieve key object by identifier",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keys"
                ],
                "summary": "Get key by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Store Identifier",
                        "name": "storeName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Key identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Key data",
                        "schema": {
                            "$ref": "#/definitions/types.KeyResponse"
                        }
                    },
                    "404": {
                        "description": "Store/Key not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Key with a specific Curve and Signing algorithm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keys"
                ],
                "summary": "Create key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Key Identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Store Identifier",
                        "name": "storeName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create key request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Key data",
                        "schema": {
                            "$ref": "#/definitions/types.KeyResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Store not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a Key by ID. The key can be recovered",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keys"
                ],
                "summary": "Soft-deletes Key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Store Identifier",
                        "name": "storeName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Key identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "404": {
                        "description": "Store/Key not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates the tags of a key by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keys"
                ],
                "summary": "Updates a key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Store Identifier",
                        "name": "storeName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Key identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Key data",
                        "schema": {
                            "$ref": "#/definitions/types.KeyResponse"
                        }
                    },
                    "404": {
                        "description": "Store/Key not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stores/{storeName}/keys/{id}/destroy": {
            "delete": {
                "description": "Permanently deletes a Key by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keys"
                ],
                "summary": "Destroys a Key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Store Identifier",
                        "name": "storeName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Key identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Destroyed successfully"
                    },
                    "404": {
                        "description": "Store/Key not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stores/{storeName}/keys/{id}/import": {
            "post": {
                "description": "Import Key with a specific Curve and Signing algorithm",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keys"
                ],
                "summary": "Import Key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Key Identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Store Identifier",
                        "name": "storeName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create key request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ImportKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Key data",
                        "schema": {
                            "$ref": "#/definitions/types.KeyResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Store not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stores/{storeName}/keys/{id}/restore": {
            "put": {
                "description": "Restores a previously soft-deleted key by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keys"
                ],
                "summary": "Restores a soft-deleted key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Store Identifier",
                        "name": "storeName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Key identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Restored successfully"
                    },
                    "404": {
                        "description": "Store/Key not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stores/{storeName}/keys/{id}/sign": {
            "post": {
                "description": "Sign random payload using a selected key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keys"
                ],
                "summary": "Sign random payload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Store Identifier",
                        "name": "storeName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Key identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Signing request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.SignBase64PayloadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "signature in base64",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Store/Key not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "IR001"
                },
                "message": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "types.KeyResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2020-07-09T12:35:42.115395Z"
                },
                "curve": {
                    "type": "string",
                    "example": "secp256k1"
                },
                "deletedAt": {
                    "type": "string",
                    "example": "2020-07-09T12:35:42.115395Z"
                },
                "destroyedAt": {
                    "type": "string",
                    "example": "2020-07-09T12:35:42.115395Z"
                },
                "disabled": {
                    "type": "boolean",
                    "example": false
                },
                "expireAt": {
                    "type": "string",
                    "example": "2020-07-09T12:35:42.115395Z"
                },
                "id": {
                    "type": "string",
                    "example": "my-key"
                },
                "publicKey": {
                    "type": "string",
                    "example": "BFVSFJhqUh9DQJwcayNtsWdD2..."
                },
                "signingAlgorithm": {
                    "type": "string",
                    "example": "ecdsa"
                },
                "tags": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2020-07-09T12:35:42.115395Z"
                }
            }
        },
...
        "types.SignBase64PayloadRequest": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "string",
                    "example": "BFVSFJhqUh9DQJwcayNtsWdD2..."
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        }
    },
...
}